local files removed and serializers folder consolidated
