+++ TODO.md
@@ -77,6 +77,7 @@
 
 #### improve deps / reqs
 - double check everything looks good
+- combine some files again
 -
 - django dans api toolkit
     - add dep for it
@@ -100,6 +101,7 @@
     - Improved and added API tests
     - All Python files now have tests
     - 90% coverage!!!
+- File structure cleaned up
 #### TODO
 
 -------------------------------------------------------
