+++ django_dans_notifications/serializers.py
@@ -1,13 +1,31 @@
+from django_dans_api_toolkit import BaseSerializer
 from rest_framework import serializers
 
-from .base import BaseSerializer
-from ..models.email import NotificationEmail, NotificationEmailTemplate
+from .models.push import NotificationPush
+from .models.email import NotificationEmailTemplate, NotificationEmail
+from .models.basic import NotificationBasic
 
-"""
-============================================================================================ #
-EMAIL SERIALIZERS ========================================================================== #
-============================================================================================ #
-"""
+
+#
+# NOTIFICATION BASIC ========================== #
+#
+class NotificationBasicSerializer(BaseSerializer):
+    class Meta:
+        model = NotificationBasic
+        fields = (
+            "id",
+            "message",
+            "read",
+            "datetime_created",
+            "datetime_sent",
+            "sent_successfully",
+            "sender",
+            "recipients",
+        )
+        read_only_fields = (
+            "id",
+            "datetime_created",
+        )
 
 
 #
@@ -56,3 +74,24 @@ class NotificationEmailSerializer(BaseSerializer):
     @staticmethod
     def get_content(obj):
         return obj.template.html_to_str(obj.context)
+
+
+#
+# NOTIFICATION PUSH ========================== #
+#
+class NotificationPushSerializer(BaseSerializer):
+    class Meta:
+        model = NotificationPush
+        fields = (
+            "id",
+            "message",
+            "datetime_created",
+            "datetime_sent",
+            "sent_successfully",
+            "sender",
+            "recipients",
+        )
+        read_only_fields = (
+            "id",
+            "datetime_created",
+        )
